+++ internal/modules/user/controller.go
@@ -2,11 +2,13 @@ package user
 
 import (
 	"errors"
-	"fmt"
 	"golang-api-starter/internal/auth"
 	"golang-api-starter/internal/helper"
-	"log"
+	"golang-api-starter/internal/helper/logger/zap_log"
+	"golang-api-starter/internal/helper/utils"
+	"sync"
 	"time"
+
 	"github.com/gofiber/fiber/v2"
 )
 
@@ -20,15 +22,18 @@ func sanitise(users Users) {
 	}
 }
 
-func NewController(s *Service) Controller {
-	return Controller{s}
+func NewController(s *Service) *Controller {
+	return &Controller{s}
 }
 
+var mu sync.Mutex
 var respCode = fiber.StatusInternalServerError
 
-/* helper func for Login & Refresh funcs below */
-func SetRefreshTokenInCookie(result map[string]interface{}, c *fiber.Ctx) {
-	cfg.LoadEnvVariables()
+/* SetTokensInCookie is a helper for Login & Refresh funcs for setting the cookies in response */
+func SetTokensInCookie(result map[string]interface{}, c *fiber.Ctx) error {
+	if result["refreshToken"] == nil && result["accessToken"] == nil {
+		return logger.Errorf("missing required 'accessToken' & 'refreshToken'")
+	}
 	env := cfg.ServerConf.Env
 	refreshToken := result["refreshToken"].(string)
 	cookie := &fiber.Cookie{
@@ -36,64 +41,91 @@ func SetRefreshTokenInCookie(result map[string]interface{}, c *fiber.Ctx) {
 		Value:    refreshToken,
 		Expires:  time.Now().Add(time.Hour * 720), // 30 days
 		HTTPOnly: true,
-		Secure:   true,
+		Secure:   env == "prod",
 		Path:     "/",
 	}
-	if env == "local" {
-		cookie.Secure = false
-	}
-
 	c.Cookie(cookie)
+
+	accessToken := result["accessToken"].(string)
+	cookie = &fiber.Cookie{
+		Name:     "accessToken",
+		Value:    accessToken,
+		Expires:  time.Now().Add(time.Hour * 720), // 30 days
+		HTTPOnly: true,
+		Secure:   env == "prod",
+		Path:     "/",
+	}
+	c.Cookie(cookie)
+
 	delete(result, "refreshToken")
+	return nil
 }
 
 func (c *Controller) Get(ctx *fiber.Ctx) error {
-	fmt.Printf("user ctrl\n")
+	logger.Debugf("user ctrl\n")
 	fctx := &helper.FiberCtx{Fctx: ctx}
-	reqCtx := &helper.ReqContext{Payload: fctx}
-	paramsMap := reqCtx.Payload.GetQueryString()
+	paramsMap := helper.GetQueryString(ctx.Request().URI().QueryString())
 	results, pagination := c.service.Get(paramsMap)
 	sanitise(results)
 
 	respCode = fiber.StatusOK
-	return ctx.
-		Status(respCode).
-		JSON(map[string]interface{}{"data": results, "pagination": pagination})
+	return fctx.JsonResponse(
+		respCode,
+		map[string]interface{}{"data": results, "pagination": pagination},
+	)
 }
 
 func (c *Controller) GetById(ctx *fiber.Ctx) error {
-	fmt.Printf("user ctrl\n")
-	id := ctx.Params("id")
+	logger.Debugf("user ctrl\n")
+	fctx := &helper.FiberCtx{Fctx: ctx}
+	id := fctx.Fctx.Params("id")
 	paramsMap := map[string]interface{}{"id": id}
 	results, err := c.service.GetById(paramsMap)
 
 	if err != nil {
 		respCode = fiber.StatusNotFound
-		return ctx.
-			Status(respCode).
-			JSON(map[string]interface{}{"message": err.Error()})
+		return fctx.JsonResponse(
+			respCode,
+			map[string]interface{}{"message": err.Error()},
+		)
 	}
 	respCode = fiber.StatusOK
-	return ctx.JSON(map[string]interface{}{"data": results[0]})
+	return fctx.JsonResponse(respCode, map[string]interface{}{"data": results[0]})
 }
 
 func (c *Controller) Create(ctx *fiber.Ctx) error {
-	fmt.Printf("user ctrl create\n")
+	logger.Debugf("user ctrl create\n")
+	c.service.ctx = ctx
 	user := &User{}
 	users := []*User{}
 
 	fctx := &helper.FiberCtx{Fctx: ctx}
 	reqCtx := &helper.ReqContext{Payload: fctx}
-	userErr, _ := reqCtx.Payload.ParseJsonToStruct(user, &users)
+	if invalidJson := reqCtx.Payload.ValidateJson(); invalidJson != nil {
+		return fctx.JsonResponse(
+			respCode,
+			map[string]interface{}{"message": invalidJson.Error()},
+		)
+	}
+
+	userErr, parseErr := reqCtx.Payload.ParseJsonToStruct(user, &users)
+	if parseErr != nil {
+		return fctx.JsonResponse(
+			fiber.StatusUnprocessableEntity,
+			map[string]interface{}{"message": parseErr.Error()},
+		)
+	}
 	if userErr == nil {
 		users = append(users, user)
 	}
-	// log.Printf("userErr: %+v, usersErr: %+v\n", userErr, usersErr)
-	// for _, t := range users {
-	// 	log.Printf("users: %+v\n", t)
-	// }
 
 	for _, user := range users {
+		if validErr := helper.ValidateStruct(*user); validErr != nil {
+			return fctx.JsonResponse(
+				fiber.StatusUnprocessableEntity,
+				map[string]interface{}{"message": validErr.Error()},
+			)
+		}
 		if user.Id == nil {
 			continue
 		} else if existing, err := c.service.GetById(map[string]interface{}{
@@ -106,61 +138,80 @@ func (c *Controller) Create(ctx *fiber.Ctx) error {
 	results, httpErr := c.service.Create(users)
 	sanitise(results)
 	if httpErr.Err != nil {
-		return ctx.
-			Status(httpErr.Code).
-			JSON(map[string]interface{}{"message": httpErr.Err.Error()})
+		return fctx.JsonResponse(
+			httpErr.Code,
+			map[string]interface{}{"message": httpErr.Err.Error()},
+		)
 	}
 
 	respCode = fiber.StatusCreated
 	if userErr == nil && len(results) > 0 {
-		return ctx.
-			Status(respCode).
-			JSON(map[string]interface{}{"data": results[0]})
+		return fctx.JsonResponse(
+			respCode,
+			map[string]interface{}{"data": results[0]},
+		)
 	}
-	return ctx.
-		Status(respCode).
-		JSON(map[string]interface{}{"data": results})
+	return fctx.JsonResponse(
+		respCode,
+		map[string]interface{}{"data": results},
+	)
 }
 
 func (c *Controller) Update(ctx *fiber.Ctx) error {
-	fmt.Printf("user ctrl update\n")
+	logger.Debugf("user ctrl update\n")
 
 	user := &User{}
 	users := []*User{}
 
 	fctx := &helper.FiberCtx{Fctx: ctx}
 	reqCtx := &helper.ReqContext{Payload: fctx}
-	userErr, _ := reqCtx.Payload.ParseJsonToStruct(user, &users)
+	if invalidJson := reqCtx.Payload.ValidateJson(); invalidJson != nil {
+		return fctx.JsonResponse(
+			fiber.StatusUnprocessableEntity,
+			map[string]interface{}{"message": invalidJson.Error()},
+		)
+	}
+
+	userErr, parseErr := reqCtx.Payload.ParseJsonToStruct(user, &users)
+	if parseErr != nil {
+		return fctx.JsonResponse(
+			fiber.StatusUnprocessableEntity,
+			map[string]interface{}{"message": parseErr.Error()},
+		)
+	}
 	if userErr == nil {
 		users = append(users, user)
 	}
-	// log.Printf("userErr: %+v, usersErr: %+v\n", userErr, usersErr)
-	// for _, t := range users {
-	// 	log.Printf("users: %+v\n", t)
-	// }
 
 	for _, user := range users {
+		if validErr := helper.ValidateStruct(*user); validErr != nil {
+			return fctx.JsonResponse(
+				fiber.StatusUnprocessableEntity,
+				map[string]interface{}{"message": validErr.Error()},
+			)
+		}
 		if user.Id == nil && user.MongoId == nil {
-			return ctx.
-				Status(respCode).
-				JSON(map[string]interface{}{"message": "please ensure all records with id for PATCH"})
+			return fctx.JsonResponse(
+				respCode,
+				map[string]interface{}{"message": "please ensure all records with id for PATCH"},
+			)
 		}
 
-		cfg.LoadEnvVariables()
 		conditions := map[string]interface{}{}
 		conditions["id"] = user.GetId()
 
 		existing, err := c.service.GetById(conditions)
 		if len(existing) == 0 {
 			respCode = fiber.StatusNotFound
-			return ctx.
-				Status(respCode).
-				JSON(map[string]interface{}{
+			return fctx.JsonResponse(
+				respCode,
+				map[string]interface{}{
 					"message": errors.Join(
 						errors.New("cannot update non-existing records..."),
 						err,
 					).Error(),
-				})
+				},
+			)
 		} else if user.CreatedAt == nil {
 			user.CreatedAt = existing[0].CreatedAt
 		}
@@ -168,21 +219,24 @@ func (c *Controller) Update(ctx *fiber.Ctx) error {
 
 	results, httpErr := c.service.Update(users)
 	if httpErr.Err != nil {
-		return ctx.
-			Status(httpErr.Code).
-			JSON(map[string]interface{}{"message": httpErr.Err.Error()})
+		return fctx.JsonResponse(
+			httpErr.Code,
+			map[string]interface{}{"message": httpErr.Err.Error()},
+		)
 	}
 	sanitise(results)
 
 	respCode = fiber.StatusOK
 	if userErr == nil && len(results) > 0 {
-		return ctx.
-			Status(respCode).
-			JSON(map[string]interface{}{"data": results[0]})
+		return fctx.JsonResponse(
+			respCode,
+			map[string]interface{}{"data": results[0]},
+		)
 	}
-	return ctx.
-		Status(respCode).
-		JSON(map[string]interface{}{"data": results})
+	return fctx.JsonResponse(
+		respCode,
+		map[string]interface{}{"data": results},
+	)
 }
 
 func (c *Controller) Delete(ctx *fiber.Ctx) error {
@@ -198,17 +252,19 @@ func (c *Controller) Delete(ctx *fiber.Ctx) error {
 	reqCtx := &helper.ReqContext{Payload: fctx}
 	intIdsErr, strIdsErr := reqCtx.Payload.ParseJsonToStruct(&delIds, &mongoDelIds)
 	if intIdsErr != nil && strIdsErr != nil {
-		log.Printf("failed to parse req json, %+v\n", errors.Join(intIdsErr, strIdsErr).Error())
-		return ctx.JSON(map[string]interface{}{"message": errors.Join(intIdsErr, strIdsErr).Error()})
+		logger.Errorf("failed to parse req json, %+v\n", errors.Join(intIdsErr, strIdsErr).Error())
+		return fctx.JsonResponse(respCode, map[string]interface{}{"message": errors.Join(intIdsErr, strIdsErr).Error()})
 	}
-	fmt.Printf("deletedIds: %+v, mongoIds: %+v\n", delIds, mongoDelIds)
+	if len(delIds.Ids) == 0 && len(mongoDelIds.Ids) == 0 {
+		return fctx.JsonResponse(respCode, map[string]interface{}{"message": "please check the req json like the follow: {\"ids\":[]}"})
+	}
+	logger.Debugf("deletedIds: %+v, mongoIds: %+v\n", delIds, mongoDelIds)
 
 	var (
 		results []*User
 		err     error
 	)
 
-	cfg.LoadEnvVariables()
 	if cfg.DbConf.Driver == "mongodb" {
 		results, err = c.service.Delete(mongoDelIds.Ids)
 	} else {
@@ -218,73 +274,84 @@ func (c *Controller) Delete(ctx *fiber.Ctx) error {
 	sanitise(results)
 
 	if err != nil {
-		log.Printf("failed to delete, err: %+v\n", err.Error())
+		logger.Errorf("failed to delete, err: %+v\n", err.Error())
 		respCode = fiber.StatusNotFound
-		return ctx.
-			Status(respCode).
-			JSON(map[string]interface{}{"message": err.Error()})
+		return fctx.JsonResponse(respCode, map[string]interface{}{"message": err.Error()})
 	}
 
 	respCode = fiber.StatusOK
-	return ctx.
-		Status(respCode).
-		JSON(map[string]interface{}{"data": results})
+	return fctx.JsonResponse(respCode, map[string]interface{}{"data": results})
 }
 
 func (c *Controller) Login(ctx *fiber.Ctx) error {
-	fmt.Printf("user ctrl create\n")
+	mu.Lock() // for avoid sqlite goroute race error
+	defer mu.Unlock()
+
+	logger.Debugf("user ctrl login")
 	user := &User{}
-	users := []*User{}
 
 	fctx := &helper.FiberCtx{Fctx: ctx}
 	reqCtx := &helper.ReqContext{Payload: fctx}
-	if userErr, _ := reqCtx.Payload.ParseJsonToStruct(user, &users); userErr != nil {
-		log.Printf("userErr: %+v\n", userErr)
+	if userErr, _ := reqCtx.Payload.ParseJsonToStruct(user, nil); userErr != nil {
+		logger.Errorf("userErr: %+v\n", userErr)
 	}
-	// log.Printf("login req: %+v\n", user)
+	if user.Password == nil {
+		return fctx.JsonResponse(respCode, map[string]interface{}{"message": "missing password..."})
+	}
+	// logger.Debugf("login req: %+v\n", user)
 
 	result, httpErr := c.service.Login(user)
 	if httpErr != nil {
-		return ctx.
-			Status(httpErr.Code).
-			JSON(map[string]interface{}{"message": httpErr.Err.Error()})
+		return fctx.JsonResponse(respCode, map[string]interface{}{"message": httpErr.Err.Error()})
 	}
 
-	SetRefreshTokenInCookie(result, ctx)
+	if err := SetTokensInCookie(result, ctx); err != nil {
+		return fctx.JsonResponse(respCode, map[string]interface{}{"message": err.Error()})
+	}
 	respCode = fiber.StatusOK
-	return ctx.
-		Status(respCode).
-		JSON(map[string]interface{}{"data": result})
+	return fctx.JsonResponse(respCode, map[string]interface{}{"data": result})
 }
 
 func (c *Controller) Refresh(ctx *fiber.Ctx) error {
+	fctx := &helper.FiberCtx{Fctx: ctx}
 	// Read cookie
 	cookie := ctx.Cookies("refreshToken")
 
 	refreshToken := "Bearer " + cookie
-	fmt.Printf("%s\n", refreshToken)
+	logger.Debugf("%s\n", refreshToken)
 
 	claims, err := auth.ParseJwt(refreshToken)
 	if claims["tokenType"] != "refreshToken" || err != nil {
 		respCode = fiber.StatusExpectationFailed
-		return ctx.
-			Status(respCode).
-			JSON(map[string]interface{}{"message": "Invalid Token type... please try to login again"})
+		return fctx.JsonResponse(
+			respCode,
+			map[string]interface{}{"message": "Invalid Token type... please try to login again"},
+		)
 	}
 
-	result := map[string]interface{}{}
-	cfg.LoadEnvVariables()
+	var (
+		result     = map[string]interface{}{}
+		refreshErr *helper.HttpErr
+	)
+
 	if cfg.DbConf.Driver == "mongodb" {
 		userId := claims["userId"].(string)
-		result, _ = c.service.Refresh(&User{MongoId: &userId})
+		result, refreshErr = c.service.Refresh(&User{MongoId: &userId})
 	} else {
 		userId := int64(claims["userId"].(float64))
-		result, _ = c.service.Refresh(&User{Id: &userId})
+		// result, refreshErr = c.service.Refresh(&User{Id: &userId})
+		result, refreshErr = c.service.Refresh(&User{Id: utils.ToPtr(helper.FlexInt(userId))})
 	}
-	SetRefreshTokenInCookie(result, ctx)
+	if refreshErr != nil {
+		return fctx.JsonResponse(
+			refreshErr.Code,
+			map[string]interface{}{"message": refreshErr.Err.Error()},
+		)
+	}
 
+	if err := SetTokensInCookie(result, ctx); err != nil {
+		return fctx.JsonResponse(respCode, map[string]interface{}{"message": err.Error()})
+	}
 	respCode = fiber.StatusOK
-	return ctx.
-		Status(respCode).
-		JSON(map[string]interface{}{"data": result})
+	return fctx.JsonResponse(respCode, map[string]interface{}{"data": result})
 }
